requirements.txt
asgiref==3.2.3
Django==3.0.2
django-ckeditor==5.8.0
django-crispy-forms==1.8.1
django-js-asset==1.2.2
Pillow==7.0.0
pytils==0.3
pytz==2019.3
slugify==0.0.1
sqlparse==0.3.0
